# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Game Theory Research Lab API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /papers:
    get:
      operationId: listPapers
      tags: [papers]
      summary: List all generated research papers
      responses:
        "200":
          description: List of papers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Paper"
    post:
      operationId: generatePaper
      tags: [papers]
      summary: Generate a research paper from completed experiments and claims
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperInput"
      responses:
        "201":
          description: Paper generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
        "400":
          description: Insufficient data
  /papers/{id}:
    get:
      operationId: getPaper
      tags: [papers]
      summary: Get a paper by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paper found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
        "404":
          description: Not found

  # ── Dashboard / Wow endpoints ─────────────────────────────────────────────
...
components:
  schemas:
...
    PaperInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        abstract:
          type: string
        gameIds:
          type: array
          items:
            type: integer
          description: Which games to include; empty means all completed games
    Paper:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        abstract:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, generating, complete]
        sections:
          type: ["string", "null"]
          description: JSON array of paper sections
        claimsJson:
          type: ["string", "null"]
          description: JSON array of claim IDs included
        experimentsJson:
          type: ["string", "null"]
          description: JSON array of experiment IDs included
        wordCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Dashboard schemas ─────────────────────────────────────────────────
...